two commit with color
