+++ app/src/main/AndroidManifest.xml
@@ -10,16 +10,15 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="false">
-        <activity android:name=".AuthActivity" android:theme="@style/Theme.MaterialComponents.Light.NoActionBar">
+        android:supportsRtl="false"
+        android:theme="@style/AppTheme">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity android:name=".MainActivity" android:theme="@style/AppTheme"/>
 
         <activity android:name=".AddItemActivity" android:theme="@style/AddItemTheme"/>
