--- app/src/main/res/drawable/ic_auth_icon.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="93dp"
-    android:height="93dp"
-    android:viewportWidth="93"
-    android:viewportHeight="93">
-  <path
-      android:pathData="M92.593,45.068l-0.015,-0.025 -13.805,-21.798a2.727,2.727 0,0 0,-2.302 -1.266L16.53,21.979c-0.947,0 -1.811,0.492 -2.302,1.266L0.407,45.068l-0.044,0.076a2.724,2.724 0,0 0,0.424 3.272l0.01,0.011 43.777,43.776 0.021,0.02a2.723,2.723 0,0 0,3.81 0l0.022,-0.02 43.776,-43.776 0.01,-0.01a2.725,2.725 0,0 0,0.38 -3.349zM71.555,27.428l-8.912,14.259L51.95,27.428h19.606zM57.4,43.776L35.602,43.776L46.5,29.245 57.4,43.775zM41.05,27.428L30.357,41.687l-8.912,-14.259h19.606zM16.514,29.82l8.722,13.956L7.675,43.776l8.839,-13.956zM9.302,49.225h18.96l11.3,30.26 -30.26,-30.26zM46.5,82.488L34.078,49.225h24.844L46.5,82.488zM53.439,79.485l11.3,-30.26L83.7,49.225l-30.26,30.26zM67.764,43.775l8.723,-13.955 8.839,13.956L67.764,43.776zM46.5,0a2.725,2.725 0,0 0,-2.724 2.725v10.898a2.725,2.725 0,0 0,5.449 0L49.225,2.725A2.725,2.725 0,0 0,46.5 0zM35.485,12.464l-7.707,-4.982a2.725,2.725 0,1 0,-2.958 4.576l7.707,4.982a2.725,2.725 0,1 0,2.958 -4.576zM68.99,8.29a2.725,2.725 0,0 0,-3.768 -0.808l-7.706,4.982a2.725,2.725 0,0 0,2.958 4.576l7.707,-4.982a2.725,2.725 0,0 0,0.809 -3.767z"
-      android:fillColor="#FFF"
-      android:fillType="nonZero"/>
-</vector>
