+++ app/src/main/res/values/colors.xml
@@ -7,6 +7,7 @@
     <color name="dark_sky_blue">#4a90e2</color>
     <color name="white_six">#efefef</color>
     <color name="apple_green">#7ed321</color>
-    <color name="white_two">#ffffff</color>
+    <color name="dark_gray_blue">#333f59</color>
+    <color name="white_five">#f2f2f2</color>
 
 </resources>
