+++ app/src/main/res/values/styles.xml
@@ -15,12 +15,4 @@
     </style>
 
 
-    <style name="TextStyle">
-        <item name="android:textSize">36sp</item>
-        <item name="android:fontFamily">sans-serif-medium</item>
-        <item name="android:textStyle">normal</item>
-        <item name="android:textColor">@color/white_two</item>
-    </style>
-
-
 </resources>
